Сделал задание про Модульность. CommonJS.
