+++ src/index.html
@@ -128,11 +128,7 @@
   </a>
 </template>
 
-<script src="js/pictures.js" charset="utf-8"</script>
-<script src="js/resizer.js" charset="utf-8"></script>
-<script src="js/upload.js" charset="utf-8"></script>
-<!-- <script src="js/check.js" charset="utf-8"></script> -->
-<!-- <script src="js/upload-stat.min.js" charset="utf-8"></script> -->
+<script src="js/main.js" charset="utf-8"></script>
 
 </body>
 </html>
