+++ src/js/main.js
@@ -1,1 +1,4 @@
 'use strict';
+
+require('./pictures');
+require('./upload');
